Latest changes with new user endpoints

+++ src/main/java/org/tracker/ubus/ubus/Components/Buses/BusPreference/Service/Impl/BusPreferenceService.java
@@ -3,10 +3,14 @@ package org.tracker.ubus.ubus.Components.Buses.BusPreference.Service.Impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.tracker.ubus.ubus.Components.Buses.BusPreference.DTO.Request.BusPreferenceDTO;
+import org.tracker.ubus.ubus.Components.Buses.BusPreference.DTO.Response.BusPreferenceResponse;
 import org.tracker.ubus.ubus.Components.Buses.BusPreference.Entity.BusPreference;
 import org.tracker.ubus.ubus.Components.Buses.BusPreference.Exceptions.BusPreferenceAlreadyExistsException;
 import org.tracker.ubus.ubus.Components.Buses.BusPreference.Exceptions.BusPreferenceMaximumException;
 import org.tracker.ubus.ubus.Components.Buses.BusPreference.Exceptions.BusPreferenceNotFoundException;
+import org.tracker.ubus.ubus.Components.Buses.BusPreference.Mapper.BusPreferenceMapper;
 import org.tracker.ubus.ubus.Components.Buses.BusPreference.Repository.BusPreferenceRepository;
 import org.tracker.ubus.ubus.Components.Buses.BusPreference.Service.Interface.IBusPreferenceService;
 import org.tracker.ubus.ubus.Components.Users.User.Entity.User;
@@ -15,15 +19,27 @@ import org.tracker.ubus.ubus.Configuration.Security.UserPrincipal;
 
 import java.util.Objects;
 
+@Transactional
 @Service
 public class BusPreferenceService implements IBusPreferenceService {
     @Autowired
     BusPreferenceRepository repository;
     @Override
-    public BusPreference addPreference(Route route) {
+    public BusPreferenceResponse addPreference(BusPreferenceDTO busPreferenceDTO) {
 
+       if(Objects.isNull(busPreferenceDTO)) {
+           throw new BusPreferenceNotFoundException("BusPreference not found");
+       }
+       if(busPreferenceDTO.getNewRoute() == null) {
+           throw new BusPreferenceNotFoundException("New route not found");
+       }
+       if(busPreferenceDTO.getNewRoute().isBlank())
+       {
+           throw new BusPreferenceNotFoundException("New route is empty");
+       }
+        Route newRoute = Route.valueOf(busPreferenceDTO.getNewRoute());
         User user = getCurrentUser();
-    boolean alreadyExists = repository.existsByUserAndRoute(user, route);
+    boolean alreadyExists = repository.existsByUserAndRoute(user, newRoute);
     if (alreadyExists) {
         throw  new BusPreferenceAlreadyExistsException("BusPreference is already set for user");
     }
@@ -33,18 +49,43 @@ public class BusPreferenceService implements IBusPreferenceService {
     {
         throw new BusPreferenceMaximumException("User already has maximum number of bus preferences set");
     }
-        BusPreference busPreference = new BusPreference(null,user,route);
-        return repository.save(busPreference);
+        BusPreference busPreference = new BusPreference(null,user,newRoute);
+        return new BusPreferenceMapper().toResponse(java.util.Collections.singletonList(repository.save(busPreference)));
     }
 
     @Override
-    public BusPreference editPreference(Route oldRoute, Route newRoute) {
-       deletePreference(oldRoute);
-       return addPreference(newRoute);
+    public BusPreferenceResponse editPreference(BusPreferenceDTO busPreferenceDTO) {
+        if(Objects.isNull(busPreferenceDTO)) {
+            throw new BusPreferenceNotFoundException("BusPreference not found");
+        }
+        if(busPreferenceDTO.getNewRoute() == null) {
+            throw new BusPreferenceNotFoundException("New route not found");
+        }
+        if(busPreferenceDTO.getOldRoute() == null) {
+            throw new BusPreferenceNotFoundException("Old route not found");
+        }
+        if(busPreferenceDTO.getOldRoute().equals(busPreferenceDTO.getNewRoute())) {
+            throw new BusPreferenceNotFoundException("Route is already set");
+        }
+       deletePreference(busPreferenceDTO);
+       return addPreference(busPreferenceDTO);
     }
 
     @Override
-    public void deletePreference(Route oldRoute) {
+    public void deletePreference(BusPreferenceDTO busPreferenceDTO) {
+
+        if(Objects.isNull(busPreferenceDTO)) {
+            throw new BusPreferenceNotFoundException("BusPreference not found");
+        }
+        if(busPreferenceDTO.getOldRoute() == null) {
+            throw new BusPreferenceNotFoundException("Old route not found");
+        }
+        if(busPreferenceDTO.getOldRoute().isBlank())
+        {
+            throw new BusPreferenceNotFoundException("Old route is empty");
+        }
+        Route oldRoute = Route.valueOf(busPreferenceDTO.getOldRoute());
+
         if(repository.existsByUserAndRoute(getCurrentUser(),oldRoute))
         {
             repository.delete(repository.findByUserAndRoute(getCurrentUser(), oldRoute));
@@ -52,6 +93,11 @@ public class BusPreferenceService implements IBusPreferenceService {
         else{
             throw new BusPreferenceNotFoundException("BusPreference not found");
         }
+    }
+
+    @Override
+    public BusPreferenceResponse viewPreferences() {
+        return new BusPreferenceMapper().toResponse(repository.findAllByUser(getCurrentUser()));
     }
